feat: support URLs with sitemaps
Also, abstract away external sources
